Separate CSE-related cases from main reducer
